[WIP] Puppetize nginx and hhvm configuration
One of the biggest parts of puppetizing the server setup.

The new roles and configurations cover all virtual hosts, including
location configuration, of all services provided by eclair.dwnet and
donut.dwnet, but NOT the installation/configuration of the underlying
services (such as shellinabox or mediawiki). Puppetizing these
configuration will be part of separate commits for each service in the
future.

Change-Id: I0a8f9f15ef4a3f8ae011fbc73923a07b55a277dc