Fix missing IPv6 support in nginx vhosts
All hosts are enabled to communicate with IPv6, however, no nginx vhost was
configured to do so. For go2tech.de domains, this resulted in problems when
a user connects via IPv6 to one of the hosts (as this is the only domain, which
has an AAAA record so far), if the client doesn't try IPv4 if IPv6 doesn't work.

This change enables IPv6 on all vhosts to fix this problem and to cleanup the
configuration.

Change-Id: I690799e6f4977ced914b6e6e1dc03563380d4d73