Use filebeat instead of logstash for forwarding log files from nodes
If logstash is not installed on a specific node, it doesn't make much sense
to install the whole logstash binary with all of it's features, if the data
is sent to another ELK stack. This also just duplicates the configuration.

Using filebeat (successor of logstash-forwarder) is more lightweight and only
forwards the log events to an existing logstash installation.