Update manifests and modules for use in Ubuntu 16.04
The first server was upgraded to Ubuntu 16.04 (the next LTS release),
which requires some updates to modules and manifests. The modules changed here
are simply updated to the latest available version.

The citoid and zotero services were changed to not use upstart anymore, but instead
usng system services to start and stop the respective services.