Remove gitlab and some other unused vhosts
The locations and vhosts for, e.g., git.go2tech.de are not used anymore
and should be removed. This commit also removes gitlab, which is not really
needed anymore, given that the update of MediaWiki-related code happens in
a different way without hosting all extensions as a copy somewhere. All other
needed repositories, like this one, were already replicated to github, which
is now the primary git repository for them.