Route all traffic of all sites through donut.dwnet
Instead of maintaining two different webservers on two servers, all sites
are now "hosted" on donut.dwnet. This also allows us to use one set of certificates
requested from Let's encrypt and that these certificates are renewed on one host
(as it has access to all webroots to verify ownership in ACME process).

donut.dwnet has itself and eclair.dwnet configured as PHP backend, so the traffic
of the pages are balanced to both servers instead of being handled by only one of them.
For that, most web pages web-roots are moved to the shareddata NFS file share, so that
both php backends are able to access them.